requirements.txt
PyYAML==6.0
tqdm~=4.64.0